+++ apps/remnant2toolkit/app/(builds)/_features/filters/_libs/filters/long-gun-filter.ts
@@ -10,13 +10,13 @@ export const longGunFilter = {
   defaultValue: longGunItems.map((item) => ({
     label: item.name,
     subLabel: item.linkedItems?.mod?.name ?? undefined,
-    value: item.id,
+    value: item.name,
     state: 'default',
   })) as FilterOption[],
   label: 'Long Gun',
   options: longGunItems.map((item) => ({
     label: item.name,
-    value: item.id,
+    value: item.name,
   })),
 } as const satisfies FilterDefinition;